requirement.txt
streamlit
tensorflow
PIL
Image
pygments.styles.dracula